# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type NewsSourceStatus = typeof NewsSourceStatus[keyof typeof NewsSourceStatus];


export const NewsSourceStatus = {
  configured: 'configured',
  partial: 'partial',
  unavailable: 'unavailable',
} as const;

export interface NewsSource {
  id: string;
  name: string;
  url: string;
  status: NewsSourceStatus;
}

export type NewsItemDirection = typeof NewsItemDirection[keyof typeof NewsItemDirection];


export const NewsItemDirection = {
  positive: 'positive',
  negative: 'negative',
  mixed: 'mixed',
} as const;

export interface NewsItem {
  id: string;
  source: string;
  sourceUrl: string;
  articleUrl: string;
  originalTitle: string;
  translatedTitle: string;
  summary: string;
  company: string;
  /** @nullable */
  ticker: string | null;
  direction: NewsItemDirection;
  pressure: string;
  why: string;
  risks: string;
  confidence: string;
  /** @nullable */
  publishedAt: string | null;
  readTime: string;
}

export interface NewsRefreshResponse {
  items: NewsItem[];
  /** @nullable */
  refreshedAt: string | null;
  sources: NewsSource[];
  warnings: string[];
}

export type NewsRefreshJobStatus = typeof NewsRefreshJobStatus[keyof typeof NewsRefreshJobStatus];


export const NewsRefreshJobStatus = {
  fetching: 'fetching',
  analyzing: 'analyzing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface NewsRefreshJob {
  jobId: string;
  status: NewsRefreshJobStatus;
  items: NewsItem[];
  startedAt: string;
  /** @nullable */
  refreshedAt: string | null;
  sources: NewsSource[];
  warnings: string[];
  /** @minimum 0 */
  processedSources: number;
  /** @minimum 1 */
  totalSources: number;
  /** @minimum 0 */
  cachedItems: number;
  /** @minimum 0 */
  newItems: number;
  /** @nullable */
  error: string | null;
}

export interface NewsRefreshError {
  error: string;
  warnings?: string[];
}