# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: news
    description: Stock market news aggregation and analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /news/sources:
    get:
      operationId: listNewsSources
      tags: [news]
      summary: List configured news sources
      responses:
        "200":
          description: Configured news sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsSource"
  /news/latest:
    get:
      operationId: getLatestNews
      tags: [news]
      summary: Return the latest cached news briefing
      responses:
        "200":
          description: Latest cached news briefing, which may be empty before the first refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsRefreshResponse"
  /news/refresh:
    post:
      operationId: refreshNews
      tags: [news]
      summary: Start a background refresh or return the currently active refresh
      responses:
        "200":
          description: Background refresh job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsRefreshJob"
        "502":
          description: Refresh could not be started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsRefreshError"
  /news/refresh/{jobId}:
    get:
      operationId: getNewsRefreshStatus
      tags: [news]
      summary: Return current progress and available results for a background refresh
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current background refresh status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsRefreshJob"
        "404":
          description: Refresh job was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsRefreshError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    NewsSource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
        status:
          type: string
          enum: [configured, partial, unavailable]
      required:
        - id
        - name
        - url
        - status
    NewsItem:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        sourceUrl:
          type: string
          format: uri
        articleUrl:
          type: string
          format: uri
        originalTitle:
          type: string
        translatedTitle:
          type: string
        summary:
          type: string
        company:
          type: string
        ticker:
          type: ["string", "null"]
        direction:
          type: string
          enum: [positive, negative, mixed]
        pressure:
          type: string
        why:
          type: string
        risks:
          type: string
        confidence:
          type: string
        publishedAt:
          type: ["string", "null"]
        readTime:
          type: string
      required:
        - id
        - source
        - sourceUrl
        - articleUrl
        - originalTitle
        - translatedTitle
        - summary
        - company
        - ticker
        - direction
        - pressure
        - why
        - risks
        - confidence
        - publishedAt
        - readTime
    NewsRefreshResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        refreshedAt:
          type: ["string", "null"]
          format: date-time
        sources:
          type: array
          items:
            $ref: "#/components/schemas/NewsSource"
        warnings:
          type: array
          items:
            type: string
      required:
        - items
        - refreshedAt
        - sources
        - warnings
    NewsRefreshJob:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [fetching, analyzing, completed, failed]
        items:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        startedAt:
          type: string
          format: date-time
        refreshedAt:
          type: ["string", "null"]
          format: date-time
        sources:
          type: array
          items:
            $ref: "#/components/schemas/NewsSource"
        warnings:
          type: array
          items:
            type: string
        processedSources:
          type: integer
          minimum: 0
        totalSources:
          type: integer
          minimum: 1
        cachedItems:
          type: integer
          minimum: 0
        newItems:
          type: integer
          minimum: 0
        error:
          type: ["string", "null"]
      required:
        - jobId
        - status
        - items
        - startedAt
        - refreshedAt
        - sources
        - warnings
        - processedSources
        - totalSources
        - cachedItems
        - newItems
        - error
    NewsRefreshError:
      type: object
      properties:
        error:
          type: string
        warnings:
          type: array
          items:
            type: string
      required:
        - error