# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsItemDirection } from './newsItemDirection';

export interface NewsItem {
  id: string;
  source: string;
  sourceUrl: string;
  articleUrl: string;
  originalTitle: string;
  translatedTitle: string;
  summary: string;
  company: string;
  /** @nullable */
  ticker: string | null;
  direction: NewsItemDirection;
  pressure: string;
  why: string;
  risks: string;
  confidence: string;
  /** @nullable */
  publishedAt: string | null;
  readTime: string;
}