# lib/api-zod/src/generated/types/newsRefreshJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsItem } from './newsItem';
import type { NewsRefreshJobStatus } from './newsRefreshJobStatus';
import type { NewsSource } from './newsSource';

export interface NewsRefreshJob {
  jobId: string;
  status: NewsRefreshJobStatus;
  items: NewsItem[];
  startedAt: Date;
  /** @nullable */
  refreshedAt: Date | null;
  sources: NewsSource[];
  warnings: string[];
  /** @minimum 0 */
  processedSources: number;
  /** @minimum 1 */
  totalSources: number;
  /** @minimum 0 */
  cachedItems: number;
  /** @minimum 0 */
  newItems: number;
  /** @nullable */
  error: string | null;
}