# lib/api-zod/src/generated/types/newsRefreshResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsItem } from './newsItem';
import type { NewsSource } from './newsSource';

export interface NewsRefreshResponse {
  items: NewsItem[];
  /** @nullable */
  refreshedAt: Date | null;
  sources: NewsSource[];
  warnings: string[];
}